requirements.txt
pytest==9.0.2
pytest-playwright==0.7.2  # сразу устанавливается и pytest и playwright
allure-pytest==2.15.3
allure-python-commons~=2.15.3
playwright~=1.57.0